backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
anthropic==0.7.0
python-dotenv==1.0.0